- removed unused local variables to allow warning-free compile with -Wall  - removed some redundant variable initializations Thanks to Mikhail Teterin for pointing these out

+++ src/generic/logutl.c
@@ -462,7 +462,7 @@ Tcl_Obj *formatMessage(LogLevel * level, char *fmt,
 		    break;
 		case 't':
 #ifdef TCL_THREADS
-		    sprintf(tmpStr, "%d", Tcl_GetCurrentThread());
+		    sprintf(tmpStr, "%d", (int) Tcl_GetCurrentThread());
 		    Tcl_AppendToObj(fmsg, tmpStr, -1);
 #else
 		    /* no thread id */
